@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --hmr --open --port 8080",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -11,36 +11,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.1",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "~7.2.1",
-    "@angular/compiler": "~7.2.1",
-    "@angular/core": "~7.2.1",
-    "@angular/forms": "~7.2.1",
-    "@angular/material": "^7.2.1",
-    "@angular/material-moment-adapter": "^7.2.1",
-    "@angular/platform-browser": "~7.2.1",
-    "@angular/platform-browser-dynamic": "~7.2.1",
-    "@angular/router": "~7.2.1",
+    "@angular/animations": "^9.0.0-next.3",
+    "@angular/cdk": "^8.1.3",
+    "@angular/common": "~9.0.0-next.3",
+    "@angular/compiler": "~9.0.0-next.3",
+    "@angular/core": "~9.0.0-next.3",
+    "@angular/forms": "~9.0.0-next.3",
+    "@angular/material": "^8.1.3",
+    "@angular/material-moment-adapter": "^8.1.3",
+    "@angular/platform-browser": "~9.0.0-next.3",
+    "@angular/platform-browser-dynamic": "~9.0.0-next.3",
+    "@angular/router": "~9.0.0-next.3",
     "core-js": "^2.6.2",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "material-design-icons": "^3.0.1",
     "moment": "^2.23.0",
-    "rxjs": "~6.3.3",
+    "rxjs": "~6.5.2",
     "simple-line-icons": "^2.4.1",
     "tslib": "^1.9.0",
-    "zone.js": "~0.8.28"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.2",
-    "@angular/cli": "^7.2.2",
-    "@angular/compiler-cli": "~7.2.1",
-    "@angular/language-service": "~7.2.1",
+    "@angular-devkit/build-angular": "~0.803.0",
+    "@angular/cli": "^8.3.0",
+    "@angular/compiler-cli": "~9.0.0-next.3",
+    "@angular/language-service": "~9.0.0-next.3",
     "@types/jasmine": "~3.3.5",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~10.12.18",
-    "codelyzer": "~4.5.0",
+    "codelyzer": "^5.1.0",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.1.4",
@@ -51,6 +51,6 @@
     "protractor": "~5.4.2",
     "ts-node": "~7.0.0",
     "tslint": "~5.12.1",
-    "typescript": "^3.2.4"
+    "typescript": "3.5.3"
   }
 }
